requirements.txt
django==1.4.3
pygithub==1.10.0
django-registration==0.8
